myApp/requirements.txt
django==3.2.2
gunicorn==20.0.4
djangorestframework==3.12.4
python-decouple==3.4
psycopg2-binary==2.9.1
bcrypt==3.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
Werkzeug==1.0.1